New GPA Improvements
New variables have been put into the model to encourage a better distribution of "top" and "bottom" students. It is finding the worst and best averages on each teams, and comparing them to the average. The further these specific students are from the average, the better. This is further ensuring the proper spread of students across teams. It's important to note that with our solver and a MIP, the optimal solution might not be one that satisfies this specific constraint, as there is other more important ones on the model.